chore: move eps resolver func to common
